Enable channel before programming sync ip
The state machine of sync ip requires the channel
to be enabled before programming the addresses so that
data written on corresponding addresses is synchronized
by IP between producer and consumer.

It was working without this in single channel demo RTL as it
had the channel enabled always internally (since it was single channel)
and was not waiting on software to enable the channel, but now
with multistream scenarios where we require explicit control
to enable/disable particular channels, the RTL is updated to
enable/disable channel as per the software programming.

So likewise, we update the application layer to enable the channel
before programming the sync ip registers.

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Signed-off-by: Chintan Vadaliya <chintanv@xilinx.com>